Extract refactoring code from preprocessingsnippets
